PS12/src/backend/requirements.txt
Flask>=2.2.5
PyMuPDF>=1.22.0
python-dotenv>=1.0.0
pyttsx3>=2.90
spacy>=3.5.0
transformers>=4.30.0
torch>=2.0.0
